pyproject.toml
[project]
name = "barai-ai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-groq>=0.3.7",
    "langgraph>=0.6.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "wikipedia>=1.4.0",
]
